Separated shared code into Message, and implemented CommandMessage
