+++ internal-messaging/message/cc/Message.cc
@@ -3,6 +3,23 @@
 #include "Message.h"
 #include <stdexcept>
 
+bool number_value(const char* string) {
+    bool decimal_occurred = false;
+    const int string_len = strlen(string);
+    for(int i = 0; i < string_len; ++i) {
+        if(string[i] == '.' && !decimal_occurred){ //first decimal occurrance
+            decimal_occurred = true;
+        } else if(decimal_occurred && string[i] == '.') { //any other decimal occurance
+            return false;
+        } else if(!isdigit(string[i])) { //not a number
+            if(i != 0 || string[i] != '-'){ //if the char is not setting the number as a negative
+                return false;
+            }
+        }
+    }  
+    return true; //string is a number
+}
+
 Message::Message(){
 	this->time_created_=0;
 }
@@ -156,4 +173,133 @@ void Message::ToString(char * string, int capacity){
         string[i] = long_string[i];
     }
 
+}
+
+int Message::FlattenHeader(char* msg, int msg_size){
+    int message_size = 0;
+    char integer_string[32];
+    char long_string[64];
+
+     // Add Sender, Recipient, and Time
+    sprintf(integer_string, "%x", this->sender_);
+    message_size += strlen(integer_string) + 1;
+    if(message_size > 255){
+        throw std::invalid_argument( "Message to large" );
+    }
+    strcat(msg, integer_string); 
+    strcat(msg, "|");
+
+    sprintf(integer_string, "%x", this->recipient_);
+    message_size += strlen(integer_string) + 1;
+    if(message_size > 255){
+        throw std::invalid_argument( "Message to large" );
+    }
+    strcat(msg, integer_string);
+    strcat(msg, "|");
+    
+    sprintf(long_string, "%lx", this->time_created_);
+    message_size += strlen(long_string) + 1;
+    if(message_size > 255){
+        throw std::invalid_argument( "Message to large" );
+    }
+    strcat(msg, long_string);
+    strcat(msg, "|");
+    return msg_size;
+}
+
+int Message::BuildHeader(char* flat, int index){
+    char integer_string[32];
+    char hex_string[32];
+    memset(integer_string, 0, 32);
+    memset(hex_string, 0, 32);
+    int i = index;
+
+    // Find sender
+    while(flat[i] != '|'){
+        sprintf(integer_string, "%c", flat[i]);
+        strcat(hex_string, integer_string); 
+        i++;
+    }
+    try {
+        this->sender_ = std::stoi(hex_string,nullptr,16);
+    } catch (std::exception const &e) {}
+
+    i++;
+    memset(integer_string, 0, 32);
+    memset(hex_string, 0, 32);
+
+    // Find recipient
+    while(flat[i] != '|'){
+        sprintf(integer_string, "%c", flat[i]);
+        strcat(hex_string, integer_string);
+        i++; 
+    }
+    try {
+        this->recipient_ = std::stoi(hex_string,nullptr,16);
+    } catch (std::exception const &e) {}
+
+    i++;
+    memset(integer_string, 0, 32);
+    memset(hex_string, 0, 32);
+    
+    // Find time sent
+    while(flat[i] != '|'){
+        sprintf(integer_string, "%c", flat[i]);
+        strcat(hex_string, integer_string); 
+        i++;
+    }
+    try {
+        this->time_created_ = std::stoi(hex_string,nullptr,16);
+    } catch (std::exception const &e) {}
+    memset(hex_string, 0, 32);
+    return i;
+}
+
+int Message::BuildContents(char* flat, int index){
+    char integer_string[32];
+    char hex_string[32];
+    char long_string[64];
+    memset(integer_string, 0, 32);
+    memset(long_string, 0, 64);
+    memset(hex_string, 0, 32);
+    int i = index;
+    int key;
+    while(flat[i] != '\3'){
+        //Get key
+        while(flat[i] != '~'){
+            sprintf(integer_string, "%c", flat[i]);
+            strcat(hex_string, integer_string); 
+            i++;
+        }
+        i++;
+        try {
+            key = std::stoi(hex_string,nullptr,16);
+        } catch (std::exception const &e) {}
+        memset(long_string, 0, 64);
+        memset(hex_string, 0, 32);
+        // Get value
+        while(flat[i] != '|'){
+            sprintf(integer_string, "%c", flat[i]);
+            strcat(hex_string, integer_string); 
+            i++;
+        }
+        try {
+            if(!number_value(hex_string)){
+                std::string value = hex_string;
+                contents_.AddKeyValuePair(key, value);
+            }
+            else if(strchr(hex_string, 46)){ // If hex_string contains a '.'
+                float value = std::stof(hex_string);
+                contents_.AddKeyValuePair(key, value);
+            }
+            else{
+                int value = std::stoul(hex_string,nullptr,16);
+                contents_.AddKeyValuePair(key, value);
+            }
+        } catch (std::exception const &e) {}
+        memset(long_string, 0, 64);
+        memset(hex_string, 0, 32);
+        i++;
+    }
+    return i;
 }
